@@ -45,53 +46,55 @@
     "@types/passport": "0.4.2",
     "@types/passport-jwt": "3.0.1",
     "@types/passport-local": "1.0.32",
-    "codelyzer": "4.1.0",
+    "codelyzer": "4.2.1",
     "concurrently": "3.5.1",
     "jasmine-core": "2.9.1",
     "jasmine-spec-reporter": "4.2.1",
     "karma": "2.0.0",
     "karma-chrome-launcher": "2.2.0",
     "karma-cli": "1.0.1",
-    "karma-coverage-istanbul-reporter": "1.3.3",
+    "karma-coverage-istanbul-reporter": "1.4.2",
     "karma-jasmine": "1.1.1",
     "karma-jasmine-html-reporter": "0.2.2",
     "node-sass": "4.7.2",
-    "nodemon-webpack-plugin": "0.1.6",
-    "protractor": "5.2.2",
-    "ts-loader": "3.2.0",
-    "ts-node": "4.1.0",
+    "nodemon-webpack-plugin": "4.0.3",
+    "protractor": "5.3.0",
+    "rimraf": "2.6.2",
+    "ts-loader": "4.3.0",
+    "ts-node": "5.0.1",
     "tslint": "5.9.1",
-    "tslint-loader": "3.5.3",
-    "typescript": "2.6.2",
-    "webpack": "3.10.0",
-    "webpack-node-externals": "1.6.0"
+    "tslint-loader": "3.6.0",
+    "typescript": "2.7.2",
+    "webpack": "4.8.3",
+    "webpack-cli": "2.1.3",
+    "webpack-node-externals": "1.7.2"
   },
   "dependencies": {
-    "@angular/animations": "5.2.9",
-    "@angular/cdk": "5.2.4",
-    "@angular/common": "5.2.9",
-    "@angular/compiler": "5.2.9",
-    "@angular/core": "5.2.9",
-    "@angular/flex-layout": "5.0.0-beta.13",
-    "@angular/forms": "5.2.9",
-    "@angular/http": "5.2.9",
-    "@angular/material": "5.2.4",
-    "@angular/platform-browser": "5.2.9",
-    "@angular/platform-browser-dynamic": "5.2.9",
-    "@angular/platform-server": "5.2.9",
-    "@angular/router": "5.2.9",
+    "@angular/animations": "6.0.1",
+    "@angular/cdk": "6.0.1",
+    "@angular/common": "6.0.1",
+    "@angular/compiler": "6.0.1",
+    "@angular/core": "6.0.1",
+    "@angular/flex-layout": "6.0.0-beta.15",
+    "@angular/forms": "6.0.1",
+    "@angular/http": "6.0.1",
+    "@angular/material": "6.0.1",
+    "@angular/platform-browser": "6.0.1",
+    "@angular/platform-browser-dynamic": "6.0.1",
+    "@angular/platform-server": "6.0.1",
+    "@angular/router": "6.0.1",
     "@nestjs/common": "4.6.6",
     "@nestjs/core": "4.6.6",
     "@nestjs/microservices": "4.6.6",
     "@nestjs/testing": "4.6.6",
     "@nestjs/websockets": "4.6.6",
-    "@nguniversal/express-engine": "5.0.0-beta.6",
-    "@nguniversal/module-map-ngfactory-loader": "5.0.0-beta.6",
+    "@nguniversal/express-engine": "6.0.0",
+    "@nguniversal/module-map-ngfactory-loader": "6.0.0",
     "@types/request": "2.47.0",
     "body-parser": "1.18.2",
     "bootstrap": "3.3.7",
     "cookie-parser": "1.4.3",
-    "core-js": "2.5.3",
+    "core-js": "2.5.4",
     "express": "4.16.2",
     "hammerjs": "2.0.8",
     "helmet": "3.11.0",
